Skeleton for dynamic, successful sign up, with hardcoded response.
This introduces the flow for a new user being registered in the app. It
hardcodes the expected response from the user to exercise the logic on
the front-end application.

When the user signs up, the same log-in is data saved on the session
storage and the user is redirected to the main app.